// Master's thesis frontend/src/app/tfm/shared/models/Data.ts
import { Category } from "./Category";

export interface Data {
  allCategories:              Category[];
  allLabels:                  string[];
  allConsumers:               string[];
  allPrices:                  number[];
  allAutonomousCommunities:   string[];
  allProvinces:               string[];
  allLocations:               string[];
}